+++ CHANGELOG.md
@@ -7,6 +7,13 @@ This document contains a historical list of changes between releases. Only
 changes that impact end-user behavior are listed; changes to documentation or
 internal API changes are not present.
 
+Main (unreleased)
+-----------------
+
+### Bugfixes
+
+- Fixed issue with defaults for Beyla component not being applied correctly. (marctc)
+
 v1.0.0 (2024-04-09)
 -------------------